Move pages to separate dir; accordion basic code
