+++ frontend/src/app/components/component.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { AccordionComponent } from './accordion/accordion.component';
+import { CommonModule } from '@angular/common';
+
+@NgModule({
+  declarations: [AccordionComponent],
+  imports: [CommonModule],
+  exports: [AccordionComponent],
+})
+export class ComponentModule {}
