+++ frontend/src/app/pages/midi-visualizer/midi-visualizer.component.html
@@ -6,6 +6,8 @@
   </div>
   <button class="click-button" onclick="runDemo()">Try demo file</button>
 
+  <app-accordion></app-accordion>
+
   <div class="dropdown-select">
     <label for="boxType">Box Type: </label>
     <select id="boxType" (change)="setBoxType()" #boxType>
