+++ frontend/src/app/pages/midi-visualizer/midi-visualizer.service.ts
